+++ src/main/kotlin/com/example/hhplus_ecommerce/domain/messaging/producer/MessageProducer.kt
@@ -1,0 +1,10 @@
+package com.example.hhplus_ecommerce.domain.messaging.producer
+
+import com.example.hhplus_ecommerce.domain.messaging.producer.dto.PaymentDataMessage
+import com.example.hhplus_ecommerce.domain.messaging.producer.dto.ProductMessage
+
+interface MessageProducer {
+	fun sendProductOrderMessage(message: ProductMessage)
+
+	fun sendAfterPaymentMessage(message: PaymentDataMessage)
+}
