+++ src/main/kotlin/com/example/hhplus_ecommerce/domain/order/event/OrderEventListener.kt
@@ -1,11 +1,11 @@
 package com.example.hhplus_ecommerce.domain.order.event
 
+import com.example.hhplus_ecommerce.domain.messaging.producer.MessageProducer
+import com.example.hhplus_ecommerce.domain.messaging.producer.dto.ProductMessage
 import com.example.hhplus_ecommerce.domain.outbox.OutboxEventStatus
 import com.example.hhplus_ecommerce.domain.outbox.ProductOrderEventOutboxService
 import com.example.hhplus_ecommerce.domain.outbox.dto.ProductOrderEventOutboxDto
 import com.example.hhplus_ecommerce.domain.outbox.dto.ProductOrderEventOutboxRequestDto
-import com.example.hhplus_ecommerce.infrastructure.kafka.producer.MessageProducer
-import com.example.hhplus_ecommerce.infrastructure.kafka.producer.dto.ProductMessage
 import org.springframework.stereotype.Component
 import org.springframework.transaction.event.TransactionPhase
 import org.springframework.transaction.event.TransactionalEventListener
