+++ src/main/kotlin/com/example/hhplus_ecommerce/infrastructure/kafka/producer/KafkaProducer.kt
@@ -1,7 +1,8 @@
 package com.example.hhplus_ecommerce.infrastructure.kafka.producer
 
-import com.example.hhplus_ecommerce.infrastructure.kafka.producer.dto.PaymentDataMessage
-import com.example.hhplus_ecommerce.infrastructure.kafka.producer.dto.ProductMessage
+import com.example.hhplus_ecommerce.domain.messaging.producer.MessageProducer
+import com.example.hhplus_ecommerce.domain.messaging.producer.dto.PaymentDataMessage
+import com.example.hhplus_ecommerce.domain.messaging.producer.dto.ProductMessage
 import com.example.hhplus_ecommerce.infrastructure.kafka.topic.AFTER_PAYMENT_TOPIC
 import com.example.hhplus_ecommerce.infrastructure.kafka.topic.PRODUCT_ORDER_TOPIC
 import org.springframework.kafka.core.KafkaTemplate
