--- src/main/kotlin/com/example/hhplus_ecommerce/infrastructure/kafka/producer/MessageProducer.kt
@@ -1,10 +1,0 @@
-package com.example.hhplus_ecommerce.infrastructure.kafka.producer
-
-import com.example.hhplus_ecommerce.infrastructure.kafka.producer.dto.PaymentDataMessage
-import com.example.hhplus_ecommerce.infrastructure.kafka.producer.dto.ProductMessage
-
-interface MessageProducer {
-	fun sendProductOrderMessage(message: ProductMessage)
-
-	fun sendAfterPaymentMessage(message: PaymentDataMessage)
-}
